v.1.0.0

+++ database/migrations/2023_06_12_175150_alter_product_attribute_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         //
          Schema::table('product_attributes', function (Blueprint $table) {
 
-            $table->unsignedInteger('attribute_id')->after('id');
+            $table->unsignedBigInteger('attribute_id')->after('id');
             $table->foreign('attribute_id')->references('id')->on('attributes')->onDelete('cascade');
 
             $table->string('value')->after('attribute_id');
